Cancel pending connection if iOS receives a premature disconnection event.
On the edge of bluetooth signal range it is possible for iOS to call didDisconnectPeripheral without ever calling didConnect.

Because the react-native-ble-plx library only subscribes to peripheral.monitorDisconnect *after* it has successfully connected, this
results in a failed connection attempt which is indiscernable from a perfectly valid pending connection to a React Native client.

This commit subscribes to monitorDisconnection before calling connect and raises BleError.cancelled() in this scenario.
The new early disconnect observer is disposed of as soon as we receive a successful connection event and it's back to business as usual.